refactor: make errors in common folder
