+++ dfget/core/downloader/p2p_downloader/power_client_test.go
@@ -24,9 +24,9 @@ import (
 	"net"
 	"net/http"
 
+	"github.com/dragonflyoss/Dragonfly/common/errors"
 	"github.com/dragonflyoss/Dragonfly/common/util"
 	"github.com/dragonflyoss/Dragonfly/dfget/core/api"
-	"github.com/dragonflyoss/Dragonfly/dfget/errors"
 	"github.com/dragonflyoss/Dragonfly/dfget/types"
 
 	"github.com/go-check/check"
