+++ supernode/daemon/mgr/dfget_task_mgr.go
@@ -5,8 +5,8 @@ import (
 	"fmt"
 
 	"github.com/dragonflyoss/Dragonfly/apis/types"
+	errorType "github.com/dragonflyoss/Dragonfly/common/errors"
 	cutil "github.com/dragonflyoss/Dragonfly/common/util"
-	errorType "github.com/dragonflyoss/Dragonfly/supernode/errors"
 
 	"github.com/pkg/errors"
 )
