+++ supernode/daemon/mgr/dfget_task_mgr_test.go
@@ -20,7 +20,7 @@ import (
 	"context"
 
 	"github.com/dragonflyoss/Dragonfly/apis/types"
-	"github.com/dragonflyoss/Dragonfly/supernode/errors"
+	"github.com/dragonflyoss/Dragonfly/common/errors"
 
 	"github.com/go-check/check"
 )
