+++ supernode/daemon/mgr/peer_mgr_test.go
@@ -21,7 +21,7 @@ import (
 	"time"
 
 	"github.com/dragonflyoss/Dragonfly/apis/types"
-	"github.com/dragonflyoss/Dragonfly/supernode/errors"
+	"github.com/dragonflyoss/Dragonfly/common/errors"
 
 	"github.com/go-check/check"
 )
